Add metadata files for crossbeam-utils.
Tests are disabled because the parser on device is considering them
failing. This is because every test that is printing unrecognized output
is considered failing. In case of this lib it's validating if threads
are behaving correctly when panicking, and panic is printing stuff to
stderr. Interestingly this is and issue only on device because there
we're parsing both stdout and stderr while as opposed to stdout only on
host.

I'll enable tests for both host and device once the parser issue is
solved.

Bug: 155309706
Change-Id: Ic7762ed7bce403d41ecb28ba4b578de614309b6a